+++ src/encoding/ManapiURL.cpp
@@ -65,8 +65,8 @@ void manapi::encoding::decode_url(std::string &dest, std::string_view str) {
                 dest += str[i];
                 continue;
             }
-            dest.push_back(static_cast<char> (manapi::unicode::onehex2dec(str[i+1]) << 4 | manapi::unicode::onehex2dec(
-                                     str[i+2])));
+            dest.push_back(static_cast<char> (manapi::unicode::onehex2dec(static_cast<uint8_t>(str[i+1])) << 4 | manapi::unicode::onehex2dec(
+                                     static_cast<uint8_t>(str[i+2]))));
             i = i + 2;
         }
     }
